Dummy implementation for linux
