--- libs/mixer---linux/melody.cpp
@@ -1,12 +1,0 @@
-#include "pxt.h"
-
-namespace music {
-//%
-void forceOutput(int outp) {
-}
-
-//%
-void playInstructions(Buffer buf) {
-}
-
-} // namespace music
