+++ libs/mixer---linux/pxt.json
@@ -1,4 +1,4 @@
 {
     "name": "mixer---linux",
-    "additionalFilePath": "../mixer"
+    "additionalFilePath": "../mixer---none"
 }
