+++ libs/mixer---none/pxt.json
@@ -1,0 +1,4 @@
+{
+    "name": "mixer---none",
+    "additionalFilePath": "../mixer"
+}
